[Feat] coredata CRUD 로직
